create and populate database from json file
