+++ src/main/java/by/exadel/internship/entity/location/City.java
@@ -1,0 +1,54 @@
+package by.exadel.internship.entity.location;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.GenericGenerator;
+
+import javax.persistence.*;
+import java.util.Objects;
+import java.util.UUID;
+
+@Entity
+@Data
+@Table(name = "city")
+@NoArgsConstructor
+@AllArgsConstructor
+public class City implements Comparable<City>{
+
+    @Id
+    @GeneratedValue(generator = "UUID")
+    @GenericGenerator(
+            name = "UUID",
+            strategy = "org.hibernate.id.UUIDGenerator"
+    )
+    @Column(name = "city_id")
+    private UUID id;
+
+    @Column(name = "city_name")
+    private String name;
+
+    @ManyToOne
+    @JoinColumn(name = "city_cntr_id")
+    private Country country;
+
+    @Override
+    public int compareTo(City o) {
+        return this.getName().compareTo(o.getName());
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        City city = (City) o;
+        return Objects.equals(id, city.id) &&
+                Objects.equals(country, city.country);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, country);
+    }
+}
