+++ src/main/resources/application.yaml
@@ -20,6 +20,7 @@ spring:
       - 'classpath:db/changelog/v.0.2/sql-import/import-form-add-columns.sql'
       - 'classpath:db/changelog/v.0.3/sql-import/import-time-for-call.sql'
 
+
 server:
   servlet:
     context-path: /api
